Do not assign names to anonymous functions
Prior to this commit, JSHint attempted to guess at a name for anonymous
functions. Presumably this was an attempt to improve the context of
function location data in the generated report. In addition to
overloading the concept of the "function name" (which has a specific
meaning in JavaScript), the heuristic employed was incomplete and
susceptable to errors.

Remove this functionality and report the name of *all* anonymous
functions as "anonymous".