Automate browser build testing
Implement a Node.js server that is capable of creating a "browser
release" version of JSHint and a browser-ready version of JSHint's unit
test suite. Use PhantomJS to execute these files in a pseudo-browser
environment and stream test status to the terminal. Finally, when all
tests have run, exit with a status code reflecting the result of the
tests.